qr-code-encoder-decoder/pyproject.toml
[project]
name = "qr-code-encoder-decoder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=11.2.1",
    "pyzbar>=0.1.9",
    "qrcode>=8.2",
    "types-qrcode>=8.2.0.20250506",
]
